move showValues to the new extender
