+++ internal/view/value_extender.go
@@ -2,9 +2,13 @@ package view
 
 import (
 	"context"
+
 	"github.com/derailed/k9s/internal"
+	"github.com/derailed/k9s/internal/client"
+	"github.com/derailed/k9s/internal/model"
 	"github.com/derailed/k9s/internal/ui"
 	"github.com/derailed/tcell/v2"
+	"github.com/rs/zerolog/log"
 )
 
 // ValueExtender adds values actions to a given viewer.
@@ -41,4 +45,35 @@ func (v *ValueExtender) valuesCmd(evt *tcell.EventKey) *tcell.EventKey {
 
 func (v *ValueExtender) defaultCtx() context.Context {
 	return context.WithValue(context.Background(), internal.KeyFactory, v.App().factory)
+}
+
+func showValues(ctx context.Context, app *App, path string, gvr client.GVR) {
+	vm := model.NewValues(gvr, path)
+	if err := vm.Init(app.factory); err != nil {
+		app.Flash().Errf("Initializing the values model failed: %s", err)
+	}
+
+	toggleValuesCmd := func(evt *tcell.EventKey) *tcell.EventKey {
+		if err := vm.ToggleValues(); err != nil {
+			app.Flash().Errf("Values toggle failed: %s", err)
+			return nil
+		}
+
+		if err := vm.Refresh(ctx); err != nil {
+			log.Error().Err(err).Msgf("values refresh failed")
+			return nil
+		}
+
+		app.Flash().Infof("Values toggled")
+		return nil
+	}
+
+	v := NewLiveView(app, "Values", vm)
+	v.actions.Add(ui.KeyActions{
+		ui.KeyV: ui.NewKeyAction("Toggle All Values", toggleValuesCmd, true),
+	})
+
+	if err := v.app.inject(v, false); err != nil {
+		v.app.Flash().Err(err)
+	}
 }
